refactor: Move object to mod
