+++ src/object/object.rs
@@ -13,28 +13,15 @@
 // limitations under the License.
 
 use std::fmt::Debug;
-use std::fmt::Display;
-use std::fmt::Formatter;
 use std::io::ErrorKind;
 use std::io::Result;
 use std::ops::RangeBounds;
-use std::pin::Pin;
-use std::sync::atomic::AtomicBool;
-use std::sync::atomic::Ordering;
 use std::sync::Arc;
-use std::sync::Mutex;
-use std::task::Context;
-use std::task::Poll;
 
 use anyhow::anyhow;
 use futures::io;
 use futures::io::Cursor;
-#[cfg(feature = "serde")]
-use serde::Deserialize;
-#[cfg(feature = "serde")]
-use serde::Serialize;
 use time::Duration;
-use time::OffsetDateTime;
 
 use crate::error::new_other_object_error;
 use crate::io::BytesRead;
@@ -44,7 +31,6 @@ use crate::io_util::CompressAlgorithm;
 #[cfg(feature = "compress")]
 use crate::io_util::DecompressReader;
 use crate::io_util::SeekableReader;
-use crate::multipart::ObjectMultipart;
 use crate::ops::BytesRange;
 use crate::ops::OpCreate;
 use crate::ops::OpCreateMultipart;
@@ -61,6 +47,11 @@ use crate::path::normalize_path;
 use crate::path::validate_path;
 use crate::Accessor;
 use crate::BlockingBytesRead;
+use crate::ObjectIterator;
+use crate::ObjectMetadata;
+use crate::ObjectMode;
+use crate::ObjectMultipart;
+use crate::ObjectStreamer;
 
 /// Handler for all object related operations.
 #[derive(Clone, Debug)]
@@ -1206,479 +1197,5 @@ impl Object {
             .create_multipart(self.path(), OpCreateMultipart::new())
             .await?;
         Ok(self.to_multipart(&upload_id))
-    }
-}
-
-/// Metadata carries all object metadata.
-///
-/// # Notes
-///
-/// mode and content_length are required metadata that all services
-/// should provide during `stat` operation. But in `list` operation,
-/// a.k.a., `ObjectEntry`'s content length could be `None`.
-#[derive(Debug, Clone, Eq, PartialEq)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub struct ObjectMetadata {
-    mode: ObjectMode,
-
-    /// # NOTE
-    ///
-    /// - For `stat` operation, content_length is required to set.
-    /// - For `list` operation, content_length could be None.
-    content_length: Option<u64>,
-    content_md5: Option<String>,
-    last_modified: Option<OffsetDateTime>,
-    etag: Option<String>,
-}
-
-impl ObjectMetadata {
-    /// Create a new object metadata
-    pub fn new(mode: ObjectMode) -> Self {
-        Self {
-            mode,
-
-            content_length: None,
-            content_md5: None,
-            last_modified: None,
-            etag: None,
-        }
-    }
-
-    /// Object mode represent this object's mode.
-    pub fn mode(&self) -> ObjectMode {
-        self.mode
-    }
-
-    /// Set mode for object.
-    pub fn set_mode(&mut self, mode: ObjectMode) -> &mut Self {
-        self.mode = mode;
-        self
-    }
-
-    /// Set mode for object.
-    pub fn with_mode(mut self, mode: ObjectMode) -> Self {
-        self.mode = mode;
-        self
-    }
-
-    /// Content length of this object.
-    ///
-    /// `Content-Length` is defined by [RFC 7230](https://httpwg.org/specs/rfc7230.html#header.content-length)
-    /// Refer to [MDN Content-Length](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Length) for more information.
-    pub fn content_length(&self) -> u64 {
-        self.content_length.unwrap_or_default()
-    }
-
-    /// Set content length of this object.
-    pub fn set_content_length(&mut self, content_length: u64) -> &mut Self {
-        self.content_length = Some(content_length);
-        self
-    }
-
-    /// Set content length of this object.
-    pub fn with_content_length(mut self, content_length: u64) -> Self {
-        self.content_length = Some(content_length);
-        self
-    }
-
-    /// Content MD5 of this object.
-    ///
-    /// Content MD5 is defined by [RFC 2616](http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html).
-    /// And removed by [RFC 7231](https://www.rfc-editor.org/rfc/rfc7231).
-    ///
-    /// OpenDAL will try its best to set this value, but not guarantee this value is the md5 of content.
-    pub fn content_md5(&self) -> Option<&str> {
-        self.content_md5.as_deref()
-    }
-
-    /// Set content MD5 of this object.
-    ///
-    /// Content MD5 is defined by [RFC 2616](http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html).
-    /// And removed by [RFC 7231](https://www.rfc-editor.org/rfc/rfc7231).
-    pub fn set_content_md5(&mut self, content_md5: &str) -> &mut Self {
-        self.content_md5 = Some(content_md5.to_string());
-        self
-    }
-
-    /// Set content MD5 of this object.
-    ///
-    /// Content MD5 is defined by [RFC 2616](http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html).
-    /// And removed by [RFC 7231](https://www.rfc-editor.org/rfc/rfc7231).
-    pub fn with_content_md5(mut self, content_md5: &str) -> Self {
-        self.content_md5 = Some(content_md5.to_string());
-        self
-    }
-
-    /// Last modified of this object.
-    ///
-    /// `Last-Modified` is defined by [RFC 7232](https://httpwg.org/specs/rfc7232.html#header.last-modified)
-    /// Refer to [MDN Last-Modified](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Last-Modified) for more information.
-    ///
-    /// OpenDAL parse the raw value into [`OffsetDateTime`] for convenient.
-    pub fn last_modified(&self) -> Option<OffsetDateTime> {
-        self.last_modified
-    }
-
-    /// Set Last modified of this object.
-    ///
-    /// `Last-Modified` is defined by [RFC 7232](https://httpwg.org/specs/rfc7232.html#header.last-modified)
-    /// Refer to [MDN Last-Modified](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Last-Modified) for more information.
-    pub fn set_last_modified(&mut self, last_modified: OffsetDateTime) -> &mut Self {
-        self.last_modified = Some(last_modified);
-        self
-    }
-
-    /// Set Last modified of this object.
-    ///
-    /// `Last-Modified` is defined by [RFC 7232](https://httpwg.org/specs/rfc7232.html#header.last-modified)
-    /// Refer to [MDN Last-Modified](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Last-Modified) for more information.
-    pub fn with_last_modified(mut self, last_modified: OffsetDateTime) -> Self {
-        self.last_modified = Some(last_modified);
-        self
-    }
-
-    /// ETag of this object.
-    ///
-    /// `ETag` is defined by [RFC 7232](https://httpwg.org/specs/rfc7232.html#header.etag)
-    /// Refer to [MDN ETag](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/ETag) for more information.
-    ///
-    /// OpenDAL will return this value AS-IS like the following:
-    ///
-    /// - `"33a64df551425fcc55e4d42a148795d9f25f89d4"`
-    /// - `W/"0815"`
-    ///
-    /// `"` is part of etag.
-    pub fn etag(&self) -> Option<&str> {
-        self.etag.as_deref()
-    }
-
-    /// Set ETag of this object.
-    ///
-    /// `ETag` is defined by [RFC 7232](https://httpwg.org/specs/rfc7232.html#header.etag)
-    /// Refer to [MDN ETag](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/ETag) for more information.
-    ///
-    /// OpenDAL will return this value AS-IS like the following:
-    ///
-    /// - `"33a64df551425fcc55e4d42a148795d9f25f89d4"`
-    /// - `W/"0815"`
-    ///
-    /// `"` is part of etag, don't trim it before setting.
-    pub fn set_etag(&mut self, etag: &str) -> &mut Self {
-        self.etag = Some(etag.to_string());
-        self
-    }
-
-    /// Set ETag of this object.
-    ///
-    /// `ETag` is defined by [RFC 7232](https://httpwg.org/specs/rfc7232.html#header.etag)
-    /// Refer to [MDN ETag](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/ETag) for more information.
-    ///
-    /// OpenDAL will return this value AS-IS like the following:
-    ///
-    /// - `"33a64df551425fcc55e4d42a148795d9f25f89d4"`
-    /// - `W/"0815"`
-    ///
-    /// `"` is part of etag, don't trim it before setting.
-    pub fn with_etag(mut self, etag: &str) -> Self {
-        self.etag = Some(etag.to_string());
-        self
-    }
-}
-
-/// ObjectMode represents the corresponding object's mode.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
-pub enum ObjectMode {
-    /// FILE means the object has data to read.
-    FILE,
-    /// DIR means the object can be listed.
-    DIR,
-    /// Unknown means we don't know what we can do on thi object.
-    Unknown,
-}
-
-impl ObjectMode {
-    /// Check if this object mode is FILE.
-    pub fn is_file(self) -> bool {
-        self == ObjectMode::FILE
-    }
-    /// Check if this object mode is DIR.
-    pub fn is_dir(self) -> bool {
-        self == ObjectMode::DIR
-    }
-}
-
-impl Default for ObjectMode {
-    fn default() -> Self {
-        Self::Unknown
-    }
-}
-
-impl Display for ObjectMode {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ObjectMode::FILE => write!(f, "file"),
-            ObjectMode::DIR => write!(f, "dir"),
-            ObjectMode::Unknown => write!(f, "unknown"),
-        }
-    }
-}
-
-/// DirStream represents a stream of Dir.
-pub trait ObjectStream: futures::Stream<Item = Result<ObjectEntry>> + Unpin + Send {}
-impl<T> ObjectStream for T where T: futures::Stream<Item = Result<ObjectEntry>> + Unpin + Send {}
-
-/// ObjectStreamer is a boxed dyn [`ObjectStream`]
-pub type ObjectStreamer = Box<dyn ObjectStream>;
-
-/// ObjectIterate represents an iterator of Dir.
-pub trait ObjectIterate: Iterator<Item = Result<ObjectEntry>> {}
-impl<T> ObjectIterate for T where T: Iterator<Item = Result<ObjectEntry>> {}
-
-/// ObjectIterator is a boxed dyn [`ObjectIterate`]
-pub type ObjectIterator = Box<dyn ObjectIterate>;
-
-/// ObjectEntry is returned by [`ObjectStream`] during object list.
-///
-/// Users can check returning object entry's mode or convert into an object without overhead.
-#[derive(Debug)]
-pub struct ObjectEntry {
-    acc: Arc<dyn Accessor>,
-    path: String,
-    meta: Arc<Mutex<ObjectMetadata>>,
-
-    /// Complete means this object entry already carries all metadata
-    /// that it could have. Don't call metadata anymore.
-    ///
-    /// # Safety
-    ///
-    /// It's safe to clone this value. Because complete will only transform
-    /// into `true` and never change back.
-    ///
-    /// For the worse case, we cloned a `false` to new ObjectEntry. The cost
-    /// is an extra stat.
-    complete: AtomicBool,
-}
-
-impl Clone for ObjectEntry {
-    fn clone(&self) -> Self {
-        Self {
-            acc: self.acc.clone(),
-            path: self.path.clone(),
-            meta: self.meta.clone(),
-            // Read comments on this field.
-            complete: self.complete.load(Ordering::Relaxed).into(),
-        }
-    }
-}
-
-impl ObjectEntry {
-    /// Create a new object entry by its corresponding underlying storage.
-    pub fn new(acc: Arc<dyn Accessor>, path: &str, meta: ObjectMetadata) -> ObjectEntry {
-        debug_assert!(
-            meta.mode.is_dir() == path.ends_with('/'),
-            "mode {:?} not match with path {}",
-            meta.mode,
-            path
-        );
-
-        ObjectEntry {
-            acc,
-            path: path.to_string(),
-            meta: Arc::new(Mutex::new(meta)),
-            complete: AtomicBool::default(),
-        }
-    }
-
-    /// Set accessor for this entry.
-    pub fn set_accessor(&mut self, acc: Arc<dyn Accessor>) {
-        self.acc = acc;
-    }
-
-    /// Complete means this object entry already carries all metadata
-    /// that it could have. Don't call metadata anymore.
-    pub fn set_complete(&mut self) -> &mut Self {
-        self.complete.store(true, Ordering::Relaxed);
-        self
-    }
-
-    /// Complete means this object entry already carries all metadata
-    /// that it could have. Don't call metadata anymore.
-    pub fn with_complete(mut self) -> Self {
-        self.complete = AtomicBool::new(true);
-        self
-    }
-
-    /// Convert [`ObjectEntry`] into [`Object`].
-    ///
-    /// This function is the same with already implemented `From` trait.
-    /// This function will make our users happier to avoid writing
-    /// generic type parameter
-    pub fn into_object(self) -> Object {
-        self.into()
-    }
-
-    /// Return this object entry's object mode.
-    pub fn mode(&self) -> ObjectMode {
-        self.meta.lock().expect("lock must succeed").mode
-    }
-
-    /// Return this object entry's id.
-    ///
-    /// The same with [`Object::id()`]
-    pub fn id(&self) -> String {
-        format!("{}{}", self.acc.metadata().root(), self.path)
-    }
-
-    /// Return this object entry's path.
-    ///
-    /// The same with [`Object::path()`]
-    pub fn path(&self) -> &str {
-        &self.path
-    }
-
-    /// Set path for this entry.
-    pub fn set_path(&mut self, path: &str) {
-        #[cfg(debug_assertions)]
-        {
-            let mode = self.meta.lock().expect("lock must succeed").mode;
-            assert_eq!(
-                mode.is_dir(),
-                path.ends_with('/'),
-                "mode {mode:?} not match with path {path}",
-            );
-        }
-
-        self.path = path.to_string();
-    }
-
-    /// Return this object entry's name.
-    ///
-    /// The same with [`Object::name()`]
-    pub fn name(&self) -> &str {
-        get_basename(&self.path)
-    }
-
-    /// Fetch metadata about this object entry.
-    pub async fn metadata(&self) -> ObjectMetadata {
-        if !self.complete.load(Ordering::Relaxed) {
-            // We will ignore all errors happened during inner metadata.
-            if let Ok(meta) = self.acc.stat(self.path(), OpStat::new()).await {
-                self.set_metadata(meta);
-                self.complete.store(true, Ordering::Relaxed);
-            }
-        }
-
-        self.meta.lock().expect("lock must succeed").clone()
-    }
-
-    /// Fetch metadata about this object entry.
-    ///
-    /// The same with [`Object::blocking_metadata()`]
-    pub fn blocking_metadata(&self) -> ObjectMetadata {
-        if !self.complete.load(Ordering::Relaxed) {
-            // We will ignore all errors happened during inner metadata.
-            if let Ok(meta) = self.acc.blocking_stat(self.path(), OpStat::new()) {
-                self.set_metadata(meta);
-                self.complete.store(true, Ordering::Relaxed);
-            }
-        }
-
-        self.meta.lock().expect("lock must succeed").clone()
-    }
-
-    /// Update ObjectEntry's metadata by setting new one.
-    pub fn set_metadata(&self, meta: ObjectMetadata) -> &Self {
-        let mut guard = self.meta.lock().expect("lock must succeed");
-        *guard = meta;
-        self
-    }
-
-    /// The size of `ObjectEntry`'s corresponding object
-    ///
-    /// `content_length` is a prefetched metadata field in `ObjectEntry`.
-    pub async fn content_length(&self) -> u64 {
-        if let Some(v) = self.meta.lock().expect("lock must succeed").content_length {
-            return v;
-        }
-
-        self.metadata().await.content_length.unwrap_or_default()
-    }
-
-    /// The MD5 message digest of `ObjectEntry`'s corresponding object
-    ///
-    /// `content_md5` is a prefetched metadata field in `ObjectEntry`
-    ///
-    /// It doesn't mean this metadata field of object doesn't exist if `content_md5` is `None`.
-    /// Then you have to call `ObjectEntry::metadata()` to get the metadata you want.
-    pub async fn content_md5(&self) -> Option<String> {
-        if let Some(v) = self
-            .meta
-            .lock()
-            .expect("lock must succeed")
-            .content_md5
-            .clone()
-        {
-            return Some(v);
-        }
-
-        self.metadata().await.content_md5
-    }
-
-    /// The last modified UTC datetime of `ObjectEntry`'s corresponding object
-    ///
-    /// `last_modified` is a prefetched metadata field in `ObjectEntry`
-    ///
-    /// It doesn't mean this metadata field of object doesn't exist if `last_modified` is `None`.
-    /// Then you have to call `ObjectEntry::metadata()` to get the metadata you want.
-    pub async fn last_modified(&self) -> Option<OffsetDateTime> {
-        if let Some(v) = self.meta.lock().expect("lock must succeed").last_modified {
-            return Some(v);
-        }
-
-        self.metadata().await.last_modified
-    }
-
-    /// The ETag string of `ObjectEntry`'s corresponding object
-    ///
-    /// `etag` is a prefetched metadata field in `ObjectEntry`.
-    ///
-    /// It doesn't mean this metadata field of object doesn't exist if `etag` is `None`.
-    /// Then you have to call `ObjectEntry::metadata()` to get the metadata you want.
-    pub async fn etag(&self) -> Option<String> {
-        if let Some(v) = self.meta.lock().expect("lock must succeed").etag.clone() {
-            return Some(v);
-        }
-
-        self.metadata().await.etag
-    }
-}
-
-/// ObjectEntry can convert into object without overhead.
-impl From<ObjectEntry> for Object {
-    fn from(d: ObjectEntry) -> Self {
-        Object::new(d.acc, &d.path)
-    }
-}
-
-/// EmptyObjectStreamer that always return None.
-pub(crate) struct EmptyObjectStreamer;
-
-impl futures::Stream for EmptyObjectStreamer {
-    type Item = Result<ObjectEntry>;
-
-    fn poll_next(self: Pin<&mut Self>, _: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        Poll::Ready(None)
-    }
-}
-
-/// EmptyObjectIterator that always return None.
-pub(crate) struct EmptyObjectIterator;
-
-impl Iterator for EmptyObjectIterator {
-    type Item = Result<ObjectEntry>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        None
     }
 }
